Move Variable classes to new variables directory
Updated build system and fixed a couple of missing "grins/" in
some include statements.
