+++ src/variables/include/grins/solid_mechanics_variables.h
